Description of data set
A description of the data set was added, and the information about what each variable is was moved from the summary statistics area to the description of the dataset in the beginning of the Jupyter notebook for clarity. The description of the data set covers where the data came from and the purpose of analyzing the data set according to the information in the "Problem" section of the canvas assignment.