Final touches
Adjusted some of the phrasing to get closer to the word count, added some additional analysis to the Age Vs. Played Hours graph to make it clearer why I chose to analyze the relationship between played hours and subscription instead of using age. Fixed the capitalization in some text boxes.